requirements.txt
Flask==3.0.3
gunicorn==21.2.0
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
fer==22.5.1
tensorflow==2.15.0